+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Command: `/cs pet`. This toggles your pet. This replaces the old system where you had to fiddle with pet names and lists and active pets.
 - Command: `/cs trpclearstats`. This removes your stats from your TRP info if they are there while leaving the rest of the contents intact.
+- You can now have custom resources, such as Sanity, added to your character sheet.
+- Commands: `/cs addresource`, `/cs removeresource` and `/cs setresource`. These are used to manage custom resources.
 - Keybinds for toggling your pet and incrementing and decrementing pet HP.
 - Window for changing the addon's settings in the Interface menu.
 - Button for toggling combat on and off in the stats frame. This affects heal rolls.